+++ tests/error_origin.js
@@ -20,25 +20,24 @@ module.exports = async ({ test, equal, check }) => [
         const tr = await test("?>", async () => equal(false, true))
         check(tr.trace, (trace) => trace.includes('error_origin.js'))
     })
-    , await test("async test without await gives await hint", async () => {
-        // const inner_awaited = await (async () => await test("async test with await", Promise.reject(null)))()
-        // const non_awaited = await (async () => test("async test without await", Promise.reject(null)))()
+    , await test("async test without await also finds error origin", async () => {
 
-        // const inner_awaited = await (async () => await test("async test with await", async ()=> {
-        //     throw 'known?'
-        // }))()
+        const inner_awaited = await (async () => await test("async test with await", async () => {
+            throw 'known?'
+        }))()
         const non_awaited = await (async () => test("async test without await", async () => {
             throw 'unknown?'
         }))()
 
         const missing_async_msg = "Possible missing 'await' statement before an async test"
 
-        // fails on ubuntu 18.04.4 node 12.18.3
-        // check(inner_awaited.trace, t1 => !t1.includes(missing_async_msg))
-        // check(inner_awaited.trace, t2 => t2.includes('error_origin.js'))
+        // maybe fails on ubuntu 18.04.4 node 12.18.3? // TODO remove comment if fixed.
+
+        check(inner_awaited.trace, t1 => !t1.includes(missing_async_msg))
+        check(inner_awaited.trace, t2 => t2.includes('error_origin.js'))
         // console.log(`inner_awaited: ${JSON.stringify(inner_awaited,null,2)}\nia-trace: ${inner_awaited.trace}`)
 
-        check(non_awaited.trace, t3 => t3.includes(missing_async_msg))
-        check(non_awaited.trace, t4 => !t4.includes('error_origin.js'))
+        check(non_awaited.trace, t3 => !t3.includes(missing_async_msg))
+        check(non_awaited.trace, t4 => t4.includes('error_origin.js'))
     })
 ]