+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ## 2020.10.
 
 * **f** Add includes file filter pattern.
+* **x** Find async test origin
 
 ## 2020.10.24